Refactored to access through pycarta.hyperthought
Previously, the functionality of pycarta was wrapped
with the hyperthought-easy-api. This separates separates
the responsibility for using data read from HypertThought
from the tools used to read that data.

The basic functionality has not changed significantly.

Reading is quite slow: ~ 1 second per child node, so it
is strongly recommended to save the resulting graph as
a JSON object through the serialize/deserialize functions
exposed in pycarta.hyperthought.