Added CDD ID to metadata schema
Tristan and Jason have been adding CDD-IDs to the CAMDEN
metadata. This has lead to a fracturing of the HyperThought
schema; elements present in one dataset but not present in
others.

The intention is to be more permissive in allowing key-value
pairs not specified in the schema. This has been handled in
the `load` function.