Created Carta Agent
Created a Carta Agent that a user will create to affect interactions 
with the Carta server. The user must specify a authentication token, 
currently a value stored in the CartaAuth cookie, but this is expected 
to change shortly. The user may also specify a specific URL where Carta 
can be found; by default, this is 
https://dev.carta.contextualize.us.com.

Features
--------
1. Agent.auth property to set the authentication token.
2. Agent.auth property cannot be used to access the authenication token, 
as this would be a security issue.
3. Agent.url property sets the base Carta API URL.
4. Agent.url property can be used to get the current base Carta API URL.
5. Agent.get(endpoint) allows a GET command to be executed against the 
Carta base API URL. Users are not generally expected to use this 
function directly, but instead, use helper functions that wrap the 
individual API calls.
6. create_agent(token, url=None) should be used to create an Agent 
instead of calling Agent(auth="Carta auth token") directly.