Update Typing
Python 3.9 introduced typing constructs `dict[A, B]`
and `list[A]` and `tuple[A, ...]` for type hinting.
These are not available in earlier versions of Python.
In these earlier versions, the `typing` library
provided Dict, List, and Tuple classes to handle this
function.

This replaces these 3.9-specific constructs with the
more general classes available in the typing library.

An change to the `asyncio` API in Python 3.7 prevents
the update to work with 3.6, but this is the only thing
that appears to be broken.

At present, only Python 3.8 and higher are supported.